break modal star selector off into seperate componenet
